workspace/20250811_222617_ocr_poc/code/requirements.txt
Pillow==8.4.0
pytesseract==0.3.8
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
opencv-python==4.5.3.20210927